requirements.txt
# Backend API
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
httpx
sqlalchemy
alembic
psycopg[binary]
PyJWT
passlib

# Data processing and AI
numpy
pandas
scikit-learn
joblib
